backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
pypdf
python-dotenv
tqdm
ollama
sentence-transformers
pydantic 
transformers